separe cleanupdir fucntions

+++ lib/core/widgets/progressbar/progress_bar_controller.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 class ProgressBarController extends ChangeNotifier {
-  int maxProgress = 0;
-  int value = 0;
+  double maxProgress = 0;
+  double value = 0;
   void done() {
     value += 1;
     notifyListeners();
